requirements.txt
openai 
requests 
Pillow 
python-dotenv
diffusers
transformers
accelerate
torch